--- skellycam/core/camera/opencv/opencv_camera_run_process.py
@@ -1,322 +1,0 @@
-import logging
-import time
-
-import cv2
-import numpy as np
-
-from skellycam.core.camera.config.camera_config import CameraConfig
-from skellycam.core.camera.opencv.create_cv2_video_capture import create_cv2_video_capture
-from skellycam.core.camera.opencv.opencv_apply_config import apply_camera_configuration
-from skellycam.core.camera.opencv.opencv_get_frame import opencv_get_frame
-from skellycam.core.camera_group.camera_group_ipc import CameraGroupIPC
-from skellycam.core.camera_group.camera_orchestrator import CameraOrchestrator, CameraStatus
-from skellycam.core.ipc.pubsub.pubsub_manager import TopicTypes
-from skellycam.core.ipc.pubsub.pubsub_topics import SetShmMessage, DeviceExtractedConfigMessage, \
-    UpdateCamerasSettingsMessage, RecordingInfoMessage, RecordingFinishedMessage
-from skellycam.core.ipc.shared_memory.frame_payload_shared_memory_ring_buffer import \
-    FramePayloadSharedMemoryRingBuffer
-from skellycam.core.recorders.videos.video_recorder import VideoRecorder
-from skellycam.core.types.numpy_record_dtypes import create_frame_dtype
-from skellycam.core.types.type_overloads import CameraIdString, TopicSubscriptionQueue, WorkerStrategy
-from skellycam.utilities.wait_functions import wait_10us, wait_10ms, wait_1ms
-
-logger = logging.getLogger(__name__)
-
-
-def opencv_camera_worker_method(camera_id: CameraIdString,
-                                config: CameraConfig,
-                                ipc: CameraGroupIPC,
-                                update_camera_settings_subscription: TopicSubscriptionQueue,
-                                shm_subscription: TopicSubscriptionQueue,
-                                recording_info_subscription: TopicSubscriptionQueue,
-                                camera_worker_strategy: WorkerStrategy,
-                                ):
-    # Configure logging in the child process
-    if camera_worker_strategy == WorkerStrategy.PROCESS:
-        # Configure logging if multiprocessing (i.e. if there is a parent process)
-        from skellycam.system.logging_configuration.configure_logging import configure_logging
-        from skellycam import LOG_LEVEL
-        configure_logging(LOG_LEVEL, ws_queue=ipc.pubsub.topics[TopicTypes.LOGS].publication)
-    logger.trace(f"Camera {camera_id} worker started")
-    orchestrator: CameraOrchestrator = ipc.camera_orchestrator
-    self_status: CameraStatus = orchestrator.camera_statuses[camera_id]
-    self_status.running.value = True
-    camera_shm: FramePayloadSharedMemoryRingBuffer | None = None
-
-    (camera_shm,
-     config,
-     cv2_video_capture,
-     frame_rec_array) = setup_camera_loop(camera_shm=camera_shm,
-                                          config=config,
-                                          ipc=ipc,
-                                          self_status=self_status,
-                                          shm_subscription=shm_subscription)
-
-    try:
-        logger.debug(f"Camera {config.camera_id} frame grab loop starting...")
-        run_camera_loop(camera_shm=camera_shm,
-                        config=config,
-                        cv2_video_capture=cv2_video_capture,
-                        frame_rec_array=frame_rec_array,
-                        ipc=ipc,
-                        orchestrator=orchestrator,
-                        self_status=self_status,
-                        update_camera_settings_subscription=update_camera_settings_subscription,
-                        recording_info_subscription=recording_info_subscription)
-
-
-
-    except Exception as e:
-        self_status.signal_error()
-        logger.exception(f"Exception occurred when running Camera Process for Camera: {camera_id} - {e}")
-        ipc.kill_everything()
-        raise
-    finally:
-        logger.debug(f"Releasing camera {camera_id} `cv2.VideoCapture` and shutting down CameraProcess")
-        self_status.signal_closing()
-        ipc.should_continue = False
-        if cv2_video_capture:
-            cv2_video_capture.release()
-        camera_shm.close()
-        self_status.closed.value = True
-
-        logger.debug(f"Camera {config.camera_index} process completed")
-
-
-def run_camera_loop(camera_shm: FramePayloadSharedMemoryRingBuffer,
-                    config: CameraConfig,
-                    cv2_video_capture: cv2.VideoCapture,
-                    frame_rec_array: np.recarray,
-                    ipc: CameraGroupIPC,
-                    orchestrator: CameraOrchestrator,
-                    self_status: CameraStatus,
-                    update_camera_settings_subscription: TopicSubscriptionQueue,
-                    recording_info_subscription: TopicSubscriptionQueue):
-    video_recorder: VideoRecorder | None = None
-    try:
-        while ipc.should_continue:
-
-            video_recorder = check_for_new_recording_info(config=config,
-                                                          ipc=ipc,
-                                                          orchestrator=orchestrator,
-                                                          recording_info_subscription=recording_info_subscription,
-                                                          self_status=self_status,
-                                                          video_recorder=video_recorder)
-
-            if ipc.should_pause.value:
-                if not self_status.is_paused.value:
-                    logger.trace(f"Pausing camera {config.camera_id}...")
-                    self_status.is_paused.value = True
-                wait_1ms()
-                continue
-            self_status.is_paused.value = False
-
-            if not orchestrator.should_grab_by_id(camera_id=config.camera_id):
-                wait_10us()
-                continue
-            frame_rec_array.frame_metadata.timestamps.pre_frame_grab_ns[0] = time.perf_counter_ns()
-            self_status.grabbing_frame.value = True
-            frame_rec_array = opencv_get_frame(cap=cv2_video_capture,
-                                               frame_rec_array=frame_rec_array, )
-
-            # get `should_record` bool BEFORE unsetting 'grabbing_frame' to avoid gaps between cameras
-            should_record_frame, should_finish_recording = orchestrator.should_record_frame_number(
-                frame_number=frame_rec_array.frame_metadata.frame_number[0], )
-
-            self_status.grabbing_frame.value = False
-
-            while orchestrator.any_grabbing_frame and ipc.should_continue:
-                # Wait for all cameras to finish grabbing frames
-                wait_10us()
-
-            if should_record_frame:
-                if video_recorder is None:
-                    raise RuntimeError("Record requested before video_recorder was created.")
-                self_status.is_recording_frame.value = True
-                video_recorder.record_frame(frame=frame_rec_array)
-                self_status.is_recording_frame.value = False
-            if should_finish_recording and video_recorder is not None:
-                if not self_status.recording_in_progress.value:
-                    raise RuntimeError(
-                        f"Finish recording requested but no recording in progress for camera {config.camera_id}.")
-                if video_recorder is None or not video_recorder.any_data_saved:
-                    raise RuntimeError(
-                        f"Recording in progress but no video_recorder or no data saved for camera {config.camera_id}.")
-                logger.info(f"Camera {config.camera_id} finishing and closing video recorder...")
-                finish_recording(ipc=ipc,video_recorder=video_recorder)
-                video_recorder = None
-                self_status.recording_in_progress.value = False
-
-            camera_shm.put_frame(frame_rec_array=frame_rec_array, overwrite=True)
-            frame_rec_array, config = check_for_new_config(current_config=config,
-                                                           frame_rec_array=frame_rec_array,
-                                                           cv2_video_capture=cv2_video_capture,
-                                                           ipc=ipc,
-                                                           self_status=self_status,
-                                                           update_camera_settings_subscription=update_camera_settings_subscription)
-
-            frame_rec_array = initialize_frame_recarray(frame_rec_array=frame_rec_array)
-
-            # Last camera to increment their frame count status triggers the next frame_grab
-            self_status.frame_count.value = frame_rec_array.frame_metadata.frame_number[0]
-
-    except Exception as e:
-        self_status.signal_error()
-        logger.exception(f"Exception occurred in camera loop for Camera: {config.camera_id} - {e}")
-        ipc.kill_everything()
-        raise
-    finally:
-        self_status.running.value = False
-        logger.debug(f"Camera {config.camera_id} loop ended.")
-
-
-def finish_recording(ipc: CameraGroupIPC,
-                        video_recorder: VideoRecorder) ->  None:
-    logger.debug(f"Camera {video_recorder.camera_id} finishing recording: {video_recorder.recording_info.recording_name}")
-    frame_metadatas = video_recorder.finish_and_close()
-    ipc.pubsub.topics[TopicTypes.RECORDING_FINISHED].publish(RecordingFinishedMessage(
-        recording_info=video_recorder.recording_info,
-        frame_metadatas=frame_metadatas,
-    ))
-    video_recorder = None
-    return video_recorder
-
-
-def check_for_new_recording_info(config: CameraConfig,
-                                 ipc: CameraGroupIPC,
-                                 orchestrator: CameraOrchestrator,
-                                 recording_info_subscription: TopicSubscriptionQueue,
-                                 self_status: CameraStatus,
-                                 video_recorder: VideoRecorder | None) -> VideoRecorder | None:
-
-    if not recording_info_subscription.empty():
-        recording_info_message = recording_info_subscription.get()
-        if not isinstance(recording_info_message, RecordingInfoMessage):
-            raise RuntimeError(
-                f"Expected RecordingInfoMessage for camera {config.camera_id}, "
-                f"but received {type(recording_info_message)}"
-            )
-        recording_info = recording_info_message.recording_info
-        if video_recorder is not None:
-            logger.info(
-                f"New recording info received, closing recording: {video_recorder.recording_info.recording_name} for camera {config.camera_id}")
-            finish_recording(ipc=ipc, video_recorder=video_recorder)
-
-        logger.info(
-            f"Camera {config.camera_id} creating recorder for recording: {recording_info.recording_name}")
-        video_recorder = VideoRecorder.create(
-            recording_info=recording_info,
-            config=config,
-        )
-        self_status.recording_in_progress.value = True
-        while not orchestrator.all_cameras_recording and ipc.should_continue:
-            # Wait for all cameras to be ready to record before starting the recording
-            wait_1ms()
-    return video_recorder
-
-
-def setup_camera_loop(camera_shm: FramePayloadSharedMemoryRingBuffer | None,
-                      config: CameraConfig,
-                      ipc: CameraGroupIPC,
-                      self_status: CameraStatus,
-                      shm_subscription: TopicSubscriptionQueue) -> tuple[
-    FramePayloadSharedMemoryRingBuffer, CameraConfig, cv2.VideoCapture, np.recarray]:
-    # Create cv2.VideoCapture object
-    try:
-        cv2_video_capture, config = create_cv2_video_capture(config)
-
-        ipc.pubsub.topics[TopicTypes.EXTRACTED_CONFIG].publish(DeviceExtractedConfigMessage(extracted_config=config))
-        self_status.connected.value = True
-        logger.debug(f"Camera {config.camera_id} connected, awaiting shm message...")
-        while camera_shm is None and ipc.should_continue:
-            wait_10ms()
-            if not shm_subscription.empty():
-                shm_message: SetShmMessage = shm_subscription.get()
-                if not isinstance(shm_message, SetShmMessage):
-                    raise RuntimeError(
-                        f"Expected SetShmMessage for camera {config.camera_id}, but received {type(shm_message)}"
-                    )
-                camera_shm_dto = shm_message.camera_group_shm_dto.camera_shm_dtos[config.camera_id]
-                logger.debug(f"Creating camera shared memory for camera {config.camera_id}...")
-                camera_shm = FramePayloadSharedMemoryRingBuffer.recreate(dto=camera_shm_dto,
-                                                                         read_only=False)
-        # Ensure camera_group_shm is properly initialized before proceeding
-        if camera_shm is None or not camera_shm.valid:
-            raise RuntimeError("Failed to initialize camera_group_shm")
-        logger.success(f"Camera {config.camera_id} ready!")
-        while not ipc.all_ready_to_start and ipc.should_continue:
-            wait_10ms()
-        frame_rec_array = create_initial_frame_rec_array(config=config,
-                                                         ipc=ipc)
-
-    except Exception as e:
-        logger.exception(f"Failed to create cv2.VideoCapture for camera {config.camera_id}: {e}")
-        self_status.signal_error()
-        ipc.kill_everything()
-        raise RuntimeError(f"Could not create cv2.VideoCapture for camera {config.camera_id}") from e
-    return camera_shm, config, cv2_video_capture, frame_rec_array
-
-
-def create_initial_frame_rec_array(config: CameraConfig, ipc: CameraGroupIPC) -> np.recarray:
-    # Create initial frame record array
-    frame_dtype = create_frame_dtype(config)
-    frame_rec_array = np.recarray(1, dtype=frame_dtype)
-    # Initialize the frame metadata
-
-    frame_rec_array.frame_metadata.camera_config[0] = config.to_numpy_record_array()
-    frame_rec_array.frame_metadata.frame_number[0] = -1
-    frame_rec_array.frame_metadata.timestamps.timebase_mapping[0] = ipc.timebase_mapping.to_numpy_record_array()
-    # Initialize the image with zeros
-    image_shape = (config.resolution.height, config.resolution.width, config.color_channels)
-    frame_rec_array.image[0] = np.zeros(image_shape, dtype=np.uint8) + config.camera_index
-    return frame_rec_array
-
-
-def check_for_new_config(current_config: CameraConfig,
-                         frame_rec_array: np.recarray,
-                         cv2_video_capture: cv2.VideoCapture,
-                         ipc: CameraGroupIPC,
-                         self_status: CameraStatus,
-                         update_camera_settings_subscription) -> tuple[np.recarray, CameraConfig]:
-    if not update_camera_settings_subscription.empty():
-        logger.debug(
-            f"Camera {frame_rec_array.frame_metadata.camera_config.camera_id[0]} received update_camera_settings_subscription message")
-        update_message = update_camera_settings_subscription.get()
-        if not isinstance(update_message, UpdateCamerasSettingsMessage):
-            raise RuntimeError(
-                f"Expected UpdateCamerasSettingsMessage for camera {frame_rec_array.frame_metadata.camera_config.camera_id[0]}, "
-                f"but received {type(update_message)}"
-            )
-        if frame_rec_array.frame_metadata.camera_config.camera_id[0] in update_message.requested_configs:
-            self_status.updating.value = True
-            new_config = update_message.requested_configs[frame_rec_array.frame_metadata.camera_config.camera_id[0]]
-            extracted_config = apply_camera_configuration(cv2_vid_capture=cv2_video_capture,
-                                                          prior_config=CameraConfig.from_numpy_record_array(
-                                                              frame_rec_array.frame_metadata.camera_config),
-                                                          config=new_config, )
-            frame_rec_array.frame_metadata.camera_config[0] = extracted_config.to_numpy_record_array()
-            ipc.pubsub.topics[TopicTypes.EXTRACTED_CONFIG].publish(
-                DeviceExtractedConfigMessage(extracted_config=extracted_config))
-            self_status.updating.value = False
-            current_config = extracted_config
-
-    return frame_rec_array, current_config
-
-
-def initialize_frame_recarray(frame_rec_array: np.recarray) -> np.recarray:
-    """Initialize timestamps for a new frame"""
-    # Reset all timestamps to 0
-    frame_rec_array.frame_metadata.timestamps.pre_frame_grab_ns[0] = 0
-    frame_rec_array.frame_metadata.timestamps.post_frame_grab_ns[0] = 0
-    frame_rec_array.frame_metadata.timestamps.pre_frame_retrieve_ns[0] = 0
-    frame_rec_array.frame_metadata.timestamps.post_frame_retrieve_ns[0] = 0
-    frame_rec_array.frame_metadata.timestamps.pre_copy_to_camera_shm_ns[0] = 0
-    frame_rec_array.frame_metadata.timestamps.pre_retrieve_from_camera_shm_ns[0] = 0
-    frame_rec_array.frame_metadata.timestamps.post_retrieve_from_camera_shm_ns[0] = 0
-    frame_rec_array.frame_metadata.timestamps.pre_copy_to_multiframe_shm_ns[0] = 0
-    frame_rec_array.frame_metadata.timestamps.pre_retrieve_from_multiframe_shm_ns[0] = 0
-    frame_rec_array.frame_metadata.timestamps.post_retrieve_from_multiframe_shm_ns[0] = 0
-
-    frame_rec_array.frame_metadata.timestamps.frame_initialized_ns[0] = time.perf_counter_ns()
-
-    return frame_rec_array
