+++ skellycam/core/camera/opencv/opencv_helpers/create_initial_frame_recarray.py
@@ -1,0 +1,20 @@
+import numpy as np
+
+from skellycam.core.camera.config.camera_config import CameraConfig
+from skellycam.core.camera_group.camera_group_ipc import CameraGroupIPC
+from skellycam.core.types.numpy_record_dtypes import create_frame_dtype
+
+
+def create_initial_frame_rec_array(config: CameraConfig, ipc: CameraGroupIPC) -> np.recarray:
+    # Create initial frame record array
+    frame_dtype = create_frame_dtype(config)
+    frame_rec_array = np.recarray(1, dtype=frame_dtype)
+    # Initialize the frame metadata
+
+    frame_rec_array.frame_metadata.camera_config[0] = config.to_numpy_record_array()
+    frame_rec_array.frame_metadata.frame_number[0] = -1
+    frame_rec_array.frame_metadata.timestamps.timebase_mapping[0] = ipc.timebase_mapping.to_numpy_record_array()
+    # Initialize the image with zeros
+    image_shape = (config.resolution.height, config.resolution.width, config.color_channels)
+    frame_rec_array.image[0] = np.zeros(image_shape, dtype=np.uint8) + config.camera_index
+    return frame_rec_array
