+++ skellycam/core/camera/opencv/opencv_helpers/extract_config.py
