+++ skellycam/core/camera/opencv/opencv_helpers/opencv_apply_config.py
@@ -3,7 +3,7 @@ import logging
 import cv2
 
 from skellycam.core.camera.config.camera_config import CameraConfig
-from skellycam.core.camera.opencv.extract_config import extract_config_from_cv2_capture
+from skellycam.core.camera.opencv.opencv_helpers.extract_config import extract_config_from_cv2_capture
 from skellycam.system.diagnostics.recommend_camera_exposure_setting import get_recommended_cv2_cap_exposure, \
     ExposureModes
